feature/ab#54-gerneralize git clone
