--- webclient/app/src/app/services/AppSetupRest.js
@@ -1,10 +1,0 @@
-import axios from "axios";
-
-class AppSetupRest {
-
-    downloadApp = (data) => {
-        return axios.post(window.location.pathname + "api/setup-app", data);
-    }
-}
-
-export default AppSetupRest;
